Entregable grupal 5/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4